// models/ListReturnsRequestFilter.ts
/**
 * RMA Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * Contact: info@gemini-commerce.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ListReturnsRequestFilter {
    'searchTerms'?: Array<string>;
    'returnIds'?: Array<string>;
    'returnStatus'?: Array<string>;
    'orderGrns'?: Array<string>;
    'orderNumbers'?: Array<string>;
    'orderItemGrns'?: Array<string>;
    'orderItemSkus'?: Array<string>;
    'customerGrns'?: Array<string>;
    'customerEmails'?: Array<string>;
    'createdAtStart'?: Date;
    'createdAtEnd'?: Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "searchTerms",
            "baseName": "searchTerms",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "returnIds",
            "baseName": "returnIds",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "returnStatus",
            "baseName": "returnStatus",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "orderGrns",
            "baseName": "orderGrns",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "orderNumbers",
            "baseName": "orderNumbers",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "orderItemGrns",
            "baseName": "orderItemGrns",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "orderItemSkus",
            "baseName": "orderItemSkus",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "customerGrns",
            "baseName": "customerGrns",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "customerEmails",
            "baseName": "customerEmails",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "createdAtStart",
            "baseName": "createdAtStart",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "createdAtEnd",
            "baseName": "createdAtEnd",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return ListReturnsRequestFilter.attributeTypeMap;
    }

    public constructor() {
    }
}