// models/OrderDataSubtotal.ts
/**
 * RMA Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * Contact: info@gemini-commerce.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { OrderDataSubtotalCode } from '../models/OrderDataSubtotalCode';
import { RmaMoney } from '../models/RmaMoney';
import { HttpFile } from '../http/http';

export class OrderDataSubtotal {
    'code'?: OrderDataSubtotalCode;
    'value'?: RmaMoney;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "OrderDataSubtotalCode",
            "format": ""
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "RmaMoney",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return OrderDataSubtotal.attributeTypeMap;
    }

    public constructor() {
    }
}