// models/OrderDataTotal.ts
/**
 * RMA Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * Contact: info@gemini-commerce.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { OrderDataTotalCode } from '../models/OrderDataTotalCode';
import { RmaMoney } from '../models/RmaMoney';
import { HttpFile } from '../http/http';

export class OrderDataTotal {
    'code'?: OrderDataTotalCode;
    'value'?: RmaMoney;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "OrderDataTotalCode",
            "format": ""
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "RmaMoney",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return OrderDataTotal.attributeTypeMap;
    }

    public constructor() {
    }
}