// models/RmaApproveReturnRequest.ts
/**
 * RMA Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * Contact: info@gemini-commerce.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RmaApproveReturnRequestItem } from '../models/RmaApproveReturnRequestItem';
import { HttpFile } from '../http/http';

export class RmaApproveReturnRequest {
    'tenantId': string;
    'id': string;
    'refundShippingCost': boolean;
    'refundPaymentCost': boolean;
    'items': Array<RmaApproveReturnRequestItem>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "tenantId",
            "baseName": "tenantId",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "refundShippingCost",
            "baseName": "refundShippingCost",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "refundPaymentCost",
            "baseName": "refundPaymentCost",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<RmaApproveReturnRequestItem>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RmaApproveReturnRequest.attributeTypeMap;
    }

    public constructor() {
    }
}